requirements.txt
fastapi==0.115.11
tensorflow==2.18.0
numpy==2.0.2
Pillow==11.1.0
gdown==5.2.0
matplotlib==3.10.0
opencv-python==4.11.0.86
uvicorn==0.34.0